+++ account_vat_on_payment/account_account_view.xml
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="utf-8"?>
+<openerp>
+    <data>
+        <record id="view_account_form" model="ir.ui.view">
+            <field name="name">account.account.form</field>
+            <field name="model">account.account</field>
+            <field name="inherit_id" ref="account.view_account_form"/>
+            <field name="arch" type="xml">
+                  <field position="after" name="child_consol_ids">
+                      <field name="vat_on_payment_related_account_id"/>
+                    </field>
+            </field>
+        </record>
+    </data>
+</openerp>
