+++ account_vat_on_payment/account_invoice.py
@@ -1,0 +1,104 @@
+# -*- coding: utf-8 -*-
+##############################################################################
+#
+#    OpenERP, Open Source Management Solution
+#    Copyright (C) 2011-2012 Domsense s.r.l. (<http://www.domsense.com>).
+#    Copyright (C) 2012-2014 Agile Business Group sagl
+#    (<http://www.agilebg.com>)
+#
+#    This program is free software: you can redistribute it and/or modify
+#    it under the terms of the GNU Affero General Public License as
+#    published by the Free Software Foundation, either version 3 of the
+#    License, or (at your option) any later version.
+#
+#    This program is distributed in the hope that it will be useful,
+#    but WITHOUT ANY WARRANTY; without even the implied warranty of
+#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+#    GNU Affero General Public License for more details.
+#
+#    You should have received a copy of the GNU Affero General Public License
+#    along with this program.  If not, see <http://www.gnu.org/licenses/>.
+#
+##############################################################################
+
+from openerp.osv import orm, fields
+from tools.translate import _
+
+
+class account_invoice(orm.Model):
+
+    def _get_vat_on_payment(self, cr, uid, context=None):
+        return self.pool.get('res.users').browse(
+            cr, uid, uid, context).company_id.vat_on_payment
+
+    def finalize_invoice_move_lines(self, cr, uid, invoice_browse, move_lines):
+        """
+        Use shadow accounts for journal entry to be generated, according to
+        account and tax code related records
+        """
+        move_lines = super(account_invoice, self).finalize_invoice_move_lines(
+            cr, uid, invoice_browse, move_lines)
+        acc_pool = self.pool.get('account.account')
+        tax_code_pool = self.pool.get('account.tax.code')
+        new_move_lines = []
+        for line_tup in move_lines:
+            if invoice_browse.vat_on_payment:
+                if line_tup[2].get('account_id', False):
+                    account = acc_pool.browse(
+                        cr, uid, line_tup[2]['account_id'])
+                    if (
+                        account.type != 'receivable'
+                        and account.type != 'payable'
+                    ):
+                        if not account.vat_on_payment_related_account_id:
+                            raise orm.except_orm(
+                                _('Error'),
+                                _('''The invoice is \'VAT on payment\' but
+                                account %s does not have a related shadow
+                                account''')
+                                % account.name)
+                        line_tup[2]['real_account_id'] = line_tup[
+                            2]['account_id']
+                        line_tup[2]['account_id'] = (
+                            account.vat_on_payment_related_account_id.id)
+                if line_tup[2].get('tax_code_id', False):
+                    tax_code = tax_code_pool.browse(
+                        cr, uid, line_tup[2]['tax_code_id'])
+                    if not tax_code.vat_on_payment_related_tax_code_id:
+                        raise orm.except_orm(
+                            _('Error'),
+                            _('''The invoice is \'VAT on payment\' but
+                            tax code %s does not have a related shadow
+                            tax code''')
+                            % tax_code.name)
+                    line_tup[2]['real_tax_code_id'] = line_tup[
+                        2]['tax_code_id']
+                    line_tup[2]['tax_code_id'] = (
+                        tax_code.vat_on_payment_related_tax_code_id.id)
+            new_move_lines.append(line_tup)
+        return new_move_lines
+
+    def onchange_partner_id(
+            self, cr, uid, ids, type, partner_id, date_invoice=False,
+            payment_term=False, partner_bank_id=False, company_id=False):
+        res = super(account_invoice, self).onchange_partner_id(
+            cr, uid, ids, type, partner_id, date_invoice, payment_term,
+            partner_bank_id, company_id)
+        # default value for VAT on Payment is changed every time the
+        # customer/supplier is changed
+        partner_obj = self.pool.get("res.partner")
+        if partner_id:
+            p = partner_obj.browse(cr, uid, partner_id)
+            if p.property_account_position:
+                res['value'][
+                    'vat_on_payment'
+                    ] = p.property_account_position.default_has_vat_on_payment
+        return res
+
+    _inherit = "account.invoice"
+    _columns = {
+        'vat_on_payment': fields.boolean('Vat on payment'),
+        }
+    _defaults = {
+        'vat_on_payment': _get_vat_on_payment,
+        }
