+++ account_vat_on_payment/account_tax_code_view.xml
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="utf-8"?>
+<openerp>
+    <data>
+
+        <record id="view_tax_code_form" model="ir.ui.view">
+            <field name="name">account.tax.code.form</field>
+            <field name="model">account.tax.code</field>
+            <field name="inherit_id" ref="account.view_tax_code_form"/>
+            <field name="arch" type="xml">
+                  <field position="after" name="company_id">
+                      <field name="vat_on_payment_related_tax_code_id"/>
+                    </field>
+            </field>
+        </record>
+    </data>
+</openerp>
