+++ account_vat_on_payment/account_voucher.py
@@ -3,7 +3,7 @@
 #
 #    OpenERP, Open Source Management Solution
 #    Copyright (C) 2011-2012 Domsense s.r.l. (<http://www.domsense.com>).
-#    Copyright (C) 2012-2013 Agile Business Group sagl
+#    Copyright (C) 2012-2014 Agile Business Group sagl
 #    (<http://www.agilebg.com>)
 #
 #    This program is free software: you can redistribute it and/or modify
@@ -311,123 +311,3 @@ class account_voucher(orm.Model):
                         cr, uid, [voucher.shadow_move_id.id])
                     move_pool.unlink(cr, uid, [voucher.shadow_move_id.id])
         return res
-
-
-class account_invoice(orm.Model):
-
-    def _get_vat_on_payment(self, cr, uid, context=None):
-        return self.pool.get('res.users').browse(
-            cr, uid, uid, context).company_id.vat_on_payment
-
-    def finalize_invoice_move_lines(self, cr, uid, invoice_browse, move_lines):
-        """
-        Use shadow accounts for journal entry to be generated, according to
-        account and tax code related records
-        """
-        move_lines = super(account_invoice, self).finalize_invoice_move_lines(
-            cr, uid, invoice_browse, move_lines)
-        acc_pool = self.pool.get('account.account')
-        tax_code_pool = self.pool.get('account.tax.code')
-        new_move_lines = []
-        for line_tup in move_lines:
-            if invoice_browse.vat_on_payment:
-                if line_tup[2].get('account_id', False):
-                    account = acc_pool.browse(
-                        cr, uid, line_tup[2]['account_id'])
-                    if (
-                        account.type != 'receivable'
-                        and account.type != 'payable'
-                    ):
-                        if not account.vat_on_payment_related_account_id:
-                            raise orm.except_orm(
-                                _('Error'),
-                                _('''The invoice is \'VAT on payment\' but
-                                account %s does not have a related shadow
-                                account''')
-                                % account.name)
-                        line_tup[2]['real_account_id'] = line_tup[
-                            2]['account_id']
-                        line_tup[2]['account_id'] = (
-                            account.vat_on_payment_related_account_id.id)
-                if line_tup[2].get('tax_code_id', False):
-                    tax_code = tax_code_pool.browse(
-                        cr, uid, line_tup[2]['tax_code_id'])
-                    if not tax_code.vat_on_payment_related_tax_code_id:
-                        raise orm.except_orm(
-                            _('Error'),
-                            _('''The invoice is \'VAT on payment\' but
-                            tax code %s does not have a related shadow
-                            tax code''')
-                            % tax_code.name)
-                    line_tup[2]['real_tax_code_id'] = line_tup[
-                        2]['tax_code_id']
-                    line_tup[2]['tax_code_id'] = (
-                        tax_code.vat_on_payment_related_tax_code_id.id)
-            new_move_lines.append(line_tup)
-        return new_move_lines
-
-    def onchange_partner_id(
-            self, cr, uid, ids, type, partner_id, date_invoice=False,
-            payment_term=False, partner_bank_id=False, company_id=False):
-        res = super(account_invoice, self).onchange_partner_id(
-            cr, uid, ids, type, partner_id, date_invoice, payment_term,
-            partner_bank_id, company_id)
-        # default value for VAT on Payment is changed every time the
-        # customer/supplier is changed
-        partner_obj = self.pool.get("res.partner")
-        if partner_id:
-            p = partner_obj.browse(cr, uid, partner_id)
-            if p.property_account_position:
-                res['value'][
-                    'vat_on_payment'
-                    ] = p.property_account_position.default_has_vat_on_payment
-        return res
-
-    _inherit = "account.invoice"
-    _columns = {
-        'vat_on_payment': fields.boolean('Vat on payment'),
-        }
-    _defaults = {
-        'vat_on_payment': _get_vat_on_payment,
-        }
-
-
-class account_move_line(orm.Model):
-    _inherit = "account.move.line"
-    _columns = {
-        'real_payment_move_id': fields.many2one(
-            'account.move', 'Real payment entry'),
-        'real_account_id': fields.many2one('account.account', 'Real account'),
-        'real_tax_code_id': fields.many2one(
-            'account.tax.code', 'Real tax code'),
-        }
-
-
-class account_account(orm.Model):
-    _inherit = "account.account"
-    _columns = {
-        'vat_on_payment_related_account_id': fields.many2one(
-            'account.account', 'Shadow Account for VAT on payment',
-            help='''Related account used for real registrations on a
-            VAT on payment basis. Set the shadow account here'''),
-        }
-
-
-class account_tax_code(orm.Model):
-    _inherit = "account.tax.code"
-    _columns = {
-        'vat_on_payment_related_tax_code_id': fields.many2one(
-            'account.tax.code', 'Shadow Tax code for VAT on payment',
-            help='''Related tax code used for real registrations on a
-            VAT on payment basis. Set the shadow tax code here'''),
-        }
-
-
-class account_journal(orm.Model):
-    _inherit = "account.journal"
-    _columns = {
-        'vat_on_payment_related_journal_id': fields.many2one(
-            'account.journal', 'Shadow Journal for VAT on payment',
-            help='''Related journal used for shadow registrations on a
-            VAT on payment basis. Set the shadow journal here'''),
-        }
