+++ account_vat_on_payment/account_voucher_view.xml
@@ -1,0 +1,27 @@
+<?xml version="1.0" encoding="utf-8"?>
+<openerp>
+    <data>
+        
+        <record id="view_vendor_receipt_form_shadow_move" model="ir.ui.view">
+            <field name="name">account.voucher.receipt.form.shadow.move</field>
+            <field name="model">account.voucher</field>
+            <field name="inherit_id" ref="account_voucher.view_vendor_receipt_form"/>
+            <field name="arch" type="xml">
+                <field name="move_ids" position="after">
+                    <field name="shadow_move_id" />
+                </field>
+            </field>
+        </record>
+        
+        <record id="view_vendor_payment_form_shadow_move" model="ir.ui.view">
+            <field name="name">account.voucher.payment.form.shadow.move</field>
+            <field name="model">account.voucher</field>
+            <field name="inherit_id" ref="account_voucher.view_vendor_payment_form"/>
+            <field name="arch" type="xml">
+                <field name="move_ids" position="after">
+                    <field name="shadow_move_id" />
+                </field>
+            </field>
+        </record>
+    </data>
+</openerp>
